Update the fetcher to run with the v3 API.
We were previously running with v1, which was deprecated a while
back. Whoops.

Added some comments/docstrings, and split files up into more logical
components. It appears that at least some parts of the API are still
braindead (switching between javascript and ISO 8601 timestamps, for
example), so most of the error correction is probably still good to
go.